semantic-chunker/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
tiktoken==0.5.2
pgvector==0.2.4
numpy==1.26.2
